+++ src/app/routes/MANAGEMENT_ROUTES.tsx
@@ -2,7 +2,7 @@ import AssetsByCategory from "../../legacies/assetManagement/AssetsByCategory";
 import RegularDepositWithdrawal from "../../legacies/assetManagement/RegularDepositWithDrawalContainer";
 import DetailInformation from "../../legacies/assetManagement/RegularDepositWithDrawalContainer/detailPage/DetailInformation";
 import DetailSetting from "../../legacies/assetManagement/RegularDepositWithDrawalContainer/detailPage/DetailSetting";
-import SavingsGoal from "../../legacies/assetManagement/SavingsGoalContainer/SavingsGoal.tsx";
+import SavingsGoal from "../../pages/AssetManagement/pages/SavingGoal/SavingsGoal.tsx";
 import { PATH } from "@constants/path.ts";
 import { RouterDOM } from "@app/types/common.ts";
 import SpendingGoal from "@legacies/assetManagement/SpendingGoal";
