--- src/legacies/assetManagement/SavingsGoalContainer/goals/Personal/components/GoalCard/index.ts
@@ -1,3 +1,0 @@
-import GoalCard from "@legacies/assetManagement/SavingsGoalContainer/goals/Personal/components/GoalCard/GoalCard.tsx";
-
-export default GoalCard;
