--- src/legacies/assetManagement/SavingsGoalContainer/goals/Personal/index.ts
@@ -1,3 +1,0 @@
-import Personal from "@legacies/assetManagement/SavingsGoalContainer/goals/Personal/Personal.tsx";
-
-export default Personal;
