--- src/legacies/assetManagement/SavingsGoalContainer/goals/Saving/index.ts
@@ -1,3 +1,0 @@
-import Saving from "@legacies/assetManagement/SavingsGoalContainer/goals/Saving/Saving.tsx";
-
-export default Saving;
