--- src/legacies/assetManagement/SavingsGoalContainer/index.ts
@@ -1,3 +1,0 @@
-import SavingsGoal from "@legacies/assetManagement/SavingsGoalContainer/SavingsGoal.tsx";
-
-export default SavingsGoal;
