+++ src/legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/MonthSpendingGoal.tsx
@@ -2,7 +2,7 @@ import RoundedPaper from "@components/common/RoundedPaper.tsx";
 import { Box, IconButton, Stack, Typography } from "@mui/material";
 import filter_main from "@assets/icons/header/filter_main.svg";
 import RoundedBorderBox from "@components/common/RoundedBorderBox.tsx";
-import { getAmount } from "@legacies/assetManagement/SavingsGoalContainer/utils.ts";
+import { getAmount } from "@pages/AssetManagement/pages/SavingGoal/utils.ts";
 import ExpandMoreRoundedIcon from "@mui/icons-material/ExpandMoreRounded";
 import moment from "moment";
 import EmptySpendCard from "@legacies/assetManagement/SpendingGoal/components/MonthSpendingGoal/EmptySpendCard.tsx";
