+++ src/legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/RegularSpendingGoal.tsx
@@ -1,7 +1,7 @@
 import RoundedPaper from "@components/common/RoundedPaper.tsx";
 import { Box, IconButton, Stack } from "@mui/material";
 import filter_main from "@assets/icons/header/filter_main.svg";
-import { getAmount } from "@legacies/assetManagement/SavingsGoalContainer/utils.ts";
+import { getAmount } from "@pages/AssetManagement/pages/SavingGoal/utils.ts";
 import GoalCard from "@legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/components/GoalCard";
 import { Highlight } from "@legacies/assetManagement/SpendingGoal/components/RegularSpendingGoal/RegularSpendingGoal.styles.ts";
 import ModifyRegularSpendingGoal, {
