+++ src/pages/AssetManagement/pages/SavingGoal/components/Personal/components/GoalCard/index.ts
@@ -1,0 +1,3 @@
+import GoalCard from "@pages/AssetManagement/pages/SavingGoal/components/Personal/components/GoalCard/GoalCard.tsx";
+
+export default GoalCard;
