+++ src/pages/AssetManagement/pages/SavingGoal/components/Personal/index.ts
@@ -1,0 +1,3 @@
+import Personal from "@pages/AssetManagement/pages/SavingGoal/components/Personal/Personal.tsx";
+
+export default Personal;
