+++ src/pages/AssetManagement/pages/SavingGoal/components/Saving/index.ts
@@ -1,0 +1,3 @@
+import Saving from "@pages/AssetManagement/pages/SavingGoal/components/Saving/Saving.tsx";
+
+export default Saving;
