+++ src/pages/AssetManagement/pages/SavingGoal/index.ts
@@ -1,0 +1,3 @@
+import SavingsGoal from "@pages/AssetManagement/pages/SavingGoal/SavingsGoal.tsx";
+
+export default SavingsGoal;
